+++ src/pull-request-state-policy.ts
@@ -9,25 +9,16 @@ import {
   localReviewRequiresManualReview,
   localReviewRetryLoopStalled,
   reviewDecisionAllowsSamePrRepair,
-  hasProcessedReviewThread,
 } from "./review-handling";
 import { shouldRunLocalReview } from "./local-review";
 import {
   mergeConflictDetected,
   summarizeChecks,
 } from "./supervisor/supervisor-reporting";
-import {
-  codexConnectorNitpickOnlyReviewThreads,
-  codexConnectorMustFixReviewThreads,
-  evaluateCodexConnectorConvergencePolicy,
-  hasCodexConnectorFindingReviewComment,
-  hasCodexConnectorPrSuccessCurrentHeadObservation,
-} from "./codex-connector-review-policy";
-import { codexConnectorCurrentHeadReviewReadiness } from "./codex-connector-review-request-decision";
+import { codexConnectorMustFixReviewThreads } from "./codex-connector-review-policy";
 import {
   configuredBotReviewFollowUpState,
   configuredBotReviewThreads,
-  latestReviewCommentAuthorIsAllowedBot,
   manualReviewThreads,
   pendingBotReviewThreads,
   staleConfiguredBotReviewThreads,
@@ -43,37 +34,28 @@ import {
   SupervisorConfig,
 } from "./core/types";
 import {
-  configuredReviewProviderKinds,
-  repoExpectsConfiguredBotReview,
-  reviewProviderWaitPolicyFromConfig,
-} from "./core/review-providers";
+  configuredBotCurrentHeadSignalPending,
+  configuredReviewBotLabel,
+  copilotReviewPending,
+  determineConfiguredBotRateLimitWait,
+  determineCopilotReviewTimeout as determineCopilotReviewTimeoutForNow,
+  shouldWaitForConfiguredBotCurrentHeadQuietPeriod,
+  shouldWaitForConfiguredBotDraftSkipRearm,
+  shouldWaitForConfiguredBotInitialGracePeriod,
+  shouldWaitForConfiguredBotLatestHeadRearm,
+  shouldWaitForCopilotReviewPropagation,
+} from "./pull-request-state-current-head-policy";
+import {
+  effectiveConfiguredBotReviewThreadsForState,
+  hasConfiguredProviderSuccess,
+  hasProvenCodexConnectorStaleReviewMetadata,
+  processedCodexConnectorMustFixThreadsExhaustedRepeatBudget,
+  shouldWaitForCodexConnectorCurrentHeadReview,
+  staleSameHeadCodexWaitHasOnlyOutdatedResidue,
+} from "./pull-request-state-codex-residue-policy";
 import { displayLocalCiCommand } from "./core/config-parsing";
 import { nowIso } from "./core/utils";
 import { hasQueuedReadyPromotionPathHygieneRepair } from "./ready-promotion-path-hygiene-repair";
-import {
-  buildStaleReviewBotRemediation,
-  isProvenStaleReviewMetadataClassification,
-} from "./supervisor/stale-review-bot-remediation";
-
-const COPILOT_REVIEW_PROPAGATION_GRACE_MS = 5_000;
-const DEFAULT_CONFIGURED_BOT_SETTLED_WAIT_MS = 5_000;
-const DEFAULT_CONFIGURED_BOT_INITIAL_GRACE_WAIT_MS = 90_000;
-
-interface CopilotReviewTimeoutStatus {
-  timedOut: boolean;
-  action: SupervisorConfig["copilotReviewTimeoutAction"] | null;
-  startedAt: string | null;
-  timedOutAt: string | null;
-  reason: string | null;
-  timeoutMinutes: number | null;
-  kind: "requested_review" | "current_head_signal" | null;
-}
-
-interface ConfiguredBotRateLimitWaitStatus {
-  active: boolean;
-  observedAt: string | null;
-  waitUntil: string | null;
-}
 
 export type GitHubWaitStep =
   | "configured_bot_rate_limit_wait"
@@ -83,6 +65,9 @@ export type GitHubWaitStep =
   | "copilot_review_propagation_wait"
   | "copilot_review_requested_wait"
   | "checks_pending";
+
+export { copilotReviewArrived } from "./pull-request-state-current-head-policy";
+export { effectiveConfiguredBotReviewThreadsForState } from "./pull-request-state-codex-residue-policy";
 
 function pullRequestStateInferenceNowMs(nowMs?: number): number {
   return nowMs ?? Date.now.call(Date);
@@ -95,553 +80,13 @@ function reviewSatisfied(pr: GitHubPullRequest): boolean {
   );
 }
 
-function configuredReviewBotLabel(config: SupervisorConfig): string {
-  return reviewProviderWaitPolicyFromConfig(config).botLabel;
-}
-
-export function copilotReviewArrived(pr: GitHubPullRequest): boolean {
-  return (pr.copilotReviewState ?? "not_requested") === "arrived" || Boolean(pr.copilotReviewArrivedAt);
-}
-
-function determineConfiguredBotRateLimitWait(
-  config: SupervisorConfig,
-  pr: GitHubPullRequest,
-  nowMs: number,
-): ConfiguredBotRateLimitWaitStatus {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  const waitMinutes = config.configuredBotRateLimitWaitMinutes ?? 0;
-  if (
-    !policy.shouldApplyRateLimitCooldown ||
-    waitMinutes <= 0 ||
-    pr.isDraft ||
-    copilotReviewArrived(pr) ||
-    !pr.configuredBotRateLimitedAt
-  ) {
-    return { active: false, observedAt: null, waitUntil: null };
-  }
-
-  const observedAtMs = Date.parse(pr.configuredBotRateLimitedAt);
-  if (Number.isNaN(observedAtMs)) {
-    return { active: false, observedAt: pr.configuredBotRateLimitedAt, waitUntil: null };
-  }
-
-  const waitUntilMs = observedAtMs + waitMinutes * 60_000;
-  return {
-    active: nowMs < waitUntilMs,
-    observedAt: pr.configuredBotRateLimitedAt,
-    waitUntil: new Date(waitUntilMs).toISOString(),
-  };
-}
-
-function hasObservedCopilotRequest(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-): boolean {
-  if (!reviewProviderWaitPolicyFromConfig(config).shouldTrackRequestedState) {
-    return false;
-  }
-
-  return Boolean(record.copilot_review_requested_observed_at && record.copilot_review_requested_head_sha === pr.headRefOid);
-}
-
-function copilotReviewPending(config: SupervisorConfig, record: IssueRunRecord, pr: GitHubPullRequest): boolean {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  if (
-    !repoExpectsConfiguredBotReview(config) ||
-    !policy.shouldWaitForRequestedReviewSignal ||
-    pr.isDraft ||
-    copilotReviewArrived(pr)
-  ) {
-    return false;
-  }
-
-  if (!policy.shouldApplyRequestedReviewTimeout && pr.configuredBotRateLimitedAt) {
-    return false;
-  }
-
-  return (pr.copilotReviewState ?? "not_requested") === "requested" || hasObservedCopilotRequest(config, record, pr);
-}
-
-function copilotReviewTimeoutStart(config: SupervisorConfig, record: IssueRunRecord, pr: GitHubPullRequest): string | null {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  if (!copilotReviewPending(config, record, pr)) {
-    return null;
-  }
-
-  if (pr.copilotReviewRequestedAt) {
-    return pr.copilotReviewRequestedAt;
-  }
-
-  if (policy.shouldTrackRequestedState && hasObservedCopilotRequest(config, record, pr)) {
-    return record.copilot_review_requested_observed_at;
-  }
-
-  return null;
-}
-
 export function determineCopilotReviewTimeout(
   config: SupervisorConfig,
   record: IssueRunRecord,
   pr: GitHubPullRequest,
   nowMs = pullRequestStateInferenceNowMs(),
-): CopilotReviewTimeoutStatus {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  const empty: CopilotReviewTimeoutStatus = {
-    timedOut: false,
-    action: null,
-    startedAt: null,
-    timedOutAt: null,
-    reason: null,
-    timeoutMinutes: null,
-    kind: null,
-  };
-  const reviewRequestTimeoutEnabled = policy.shouldApplyRequestedReviewTimeout && config.copilotReviewWaitMinutes > 0;
-  const requestedReviewStartedAt = reviewRequestTimeoutEnabled ? copilotReviewTimeoutStart(config, record, pr) : null;
-  const requestedReviewTimeout: CopilotReviewTimeoutStatus = requestedReviewStartedAt
-    ? (() => {
-        const requestedAtMs = Date.parse(requestedReviewStartedAt);
-        if (!Number.isNaN(requestedAtMs)) {
-          const timeoutMs = config.copilotReviewWaitMinutes * 60_000;
-          if (nowMs >= requestedAtMs + timeoutMs) {
-            return {
-              timedOut: true,
-              action: config.copilotReviewTimeoutAction,
-              startedAt: requestedReviewStartedAt,
-              timedOutAt: new Date(requestedAtMs + timeoutMs).toISOString(),
-              reason:
-                `Requested ${configuredReviewBotLabel(config)} review never arrived within ${config.copilotReviewWaitMinutes} minute(s) ` +
-                `for head ${pr.headRefOid}.`,
-              timeoutMinutes: config.copilotReviewWaitMinutes,
-              kind: "requested_review",
-            };
-          }
-        }
-
-        return {
-          timedOut: false,
-          action: null,
-          startedAt: requestedReviewStartedAt,
-          timedOutAt: null,
-          reason: null,
-          timeoutMinutes: config.copilotReviewWaitMinutes,
-          kind: "requested_review",
-        };
-      })()
-    : empty;
-
-  const configuredBotTimeoutEnabled =
-    requiresConfiguredBotCurrentHeadSignal(config) && (config.configuredBotCurrentHeadSignalTimeoutMinutes ?? 0) > 0;
-  const currentHeadSignalStartedAt = configuredBotTimeoutEnabled
-    ? configuredBotCurrentHeadSignalTimeoutStartAt(config, record, pr)
-    : null;
-  const currentHeadSignalTimeout: CopilotReviewTimeoutStatus = currentHeadSignalStartedAt
-    ? (() => {
-        const startedAtMs = Date.parse(currentHeadSignalStartedAt);
-        if (Number.isNaN(startedAtMs)) {
-          return {
-            timedOut: false,
-            action: null,
-            startedAt: currentHeadSignalStartedAt,
-            timedOutAt: null,
-            reason: null,
-            timeoutMinutes: config.configuredBotCurrentHeadSignalTimeoutMinutes ?? null,
-            kind: "current_head_signal",
-          };
-        }
-
-        const timeoutMs = (config.configuredBotCurrentHeadSignalTimeoutMinutes ?? 0) * 60_000;
-        if (nowMs < startedAtMs + timeoutMs) {
-          return {
-            timedOut: false,
-            action: null,
-            startedAt: currentHeadSignalStartedAt,
-            timedOutAt: null,
-            reason: null,
-            timeoutMinutes: config.configuredBotCurrentHeadSignalTimeoutMinutes ?? null,
-            kind: "current_head_signal",
-          };
-        }
-
-        const timedOutAt = new Date(startedAtMs + timeoutMs).toISOString();
-        return {
-          timedOut: true,
-          action: config.configuredBotCurrentHeadSignalTimeoutAction ?? "block",
-          startedAt: currentHeadSignalStartedAt,
-          timedOutAt,
-          reason:
-            `${configuredReviewBotLabel(config)} never produced a current-head review signal within ` +
-            `${config.configuredBotCurrentHeadSignalTimeoutMinutes} minute(s) for head ${pr.headRefOid}.`,
-          timeoutMinutes: config.configuredBotCurrentHeadSignalTimeoutMinutes ?? null,
-          kind: "current_head_signal",
-        };
-      })()
-    : empty;
-
-  if (currentHeadSignalTimeout.kind !== null) {
-    return currentHeadSignalTimeout;
-  }
-
-  return requestedReviewTimeout;
-}
-
-function shouldWaitForCopilotReviewPropagation(
-  config: SupervisorConfig,
-  record: Pick<IssueRunRecord, "review_wait_started_at" | "review_wait_head_sha">,
-  pr: GitHubPullRequest,
-  nowMs: number,
-): boolean {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  if (
-    !policy.shouldWaitForRequestPropagation ||
-    config.copilotReviewWaitMinutes <= 0 ||
-    pr.isDraft ||
-    pr.configuredBotCurrentHeadObservedAt ||
-    pr.headRefOid !== record.review_wait_head_sha
-  ) {
-    return false;
-  }
-
-  const lifecycleState = pr.copilotReviewState ?? "not_requested";
-  if (lifecycleState === "requested" || lifecycleState === "arrived") {
-    return false;
-  }
-
-  const startedAt = record.review_wait_started_at;
-  if (!startedAt) {
-    return false;
-  }
-
-  const startedAtMs = Date.parse(startedAt);
-  if (Number.isNaN(startedAtMs)) {
-    return false;
-  }
-
-  return nowMs < startedAtMs + COPILOT_REVIEW_PROPAGATION_GRACE_MS;
-}
-
-function shouldWaitForConfiguredBotCurrentHeadQuietPeriod(
-  config: SupervisorConfig,
-  pr: GitHubPullRequest,
-  nowMs: number,
-): boolean {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  if (!policy.shouldApplyCurrentHeadQuietPeriod || pr.isDraft || !pr.configuredBotCurrentHeadObservedAt) {
-    return false;
-  }
-
-  const observedAtMs = Date.parse(pr.configuredBotCurrentHeadObservedAt);
-  if (Number.isNaN(observedAtMs)) {
-    return false;
-  }
-
-  const settledWaitMs = (config.configuredBotSettledWaitSeconds ?? DEFAULT_CONFIGURED_BOT_SETTLED_WAIT_MS / 1_000) * 1_000;
-  return nowMs < observedAtMs + settledWaitMs;
-}
-
-function validTimestamp(value: string | null | undefined): string | null {
-  if (typeof value !== "string" || value.length === 0) {
-    return null;
-  }
-
-  return Number.isNaN(Date.parse(value)) ? null : value;
-}
-
-function hasCurrentHeadProviderSuccess(
-  record: Pick<IssueRunRecord, "provider_success_head_sha" | "provider_success_observed_at">,
-  pr: GitHubPullRequest,
-): boolean {
-  return (
-    record.provider_success_head_sha === pr.headRefOid &&
-    validTimestamp(record.provider_success_observed_at) !== null
-  );
-}
-
-function currentHeadObservationSatisfiesActiveWait(
-  record: Pick<IssueRunRecord, "review_wait_started_at" | "review_wait_head_sha">,
-  pr: GitHubPullRequest,
-): boolean {
-  if (pr.configuredBotCurrentHeadObservationSource !== "codex_pr_success_comment") {
-    return true;
-  }
-
-  const observedAt = validTimestamp(pr.configuredBotCurrentHeadObservedAt);
-  const waitStartedAt = validTimestamp(record.review_wait_started_at);
-  if (!observedAt) {
-    return false;
-  }
-
-  if (!waitStartedAt || record.review_wait_head_sha !== pr.headRefOid) {
-    return true;
-  }
-
-  return Date.parse(observedAt) >= Date.parse(waitStartedAt);
-}
-
-function shouldWaitForConfiguredBotInitialGracePeriod(
-  config: SupervisorConfig,
-  pr: GitHubPullRequest,
-  nowMs: number,
-): boolean {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  if (
-    !policy.shouldApplyCurrentHeadQuietPeriod ||
-    pr.isDraft ||
-    pr.configuredBotCurrentHeadObservedAt ||
-    !pr.currentHeadCiGreenAt
-  ) {
-    return false;
-  }
-
-  const ciGreenAtMs = Date.parse(pr.currentHeadCiGreenAt);
-  if (Number.isNaN(ciGreenAtMs)) {
-    return false;
-  }
-
-  return nowMs < ciGreenAtMs + configuredBotInitialGraceWaitMs(config);
-}
-
-function requiresConfiguredBotCurrentHeadSignal(config: SupervisorConfig): boolean {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  return (
-    policy.shouldApplyCurrentHeadQuietPeriod &&
-    (config.configuredBotRequireCurrentHeadSignal === true || configuredReviewProviderKinds(config).includes("codex"))
-  );
-}
-
-function latestConfiguredBotActionableSignalAt(pr: GitHubPullRequest): string | null {
-  const candidates = [
-    pr.configuredBotCurrentHeadObservedAt,
-    pr.copilotReviewArrivedAt,
-    pr.configuredBotTopLevelReviewSubmittedAt,
-  ]
-    .map((value) => {
-      if (typeof value !== "string" || value.length === 0) {
-        return null;
-      }
-
-      const timestampMs = Date.parse(value);
-      if (Number.isNaN(timestampMs)) {
-        return null;
-      }
-
-      return { value, timestampMs };
-    })
-    .filter((candidate): candidate is { value: string; timestampMs: number } => candidate !== null);
-
-  if (candidates.length === 0) {
-    return null;
-  }
-
-  return candidates.reduce((latest, candidate) => (candidate.timestampMs > latest.timestampMs ? candidate : latest)).value;
-}
-
-function shouldWaitForConfiguredBotDraftSkipRearm(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-  nowMs: number,
-): boolean {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  if (!policy.shouldApplyCurrentHeadQuietPeriod || pr.isDraft || !pr.configuredBotDraftSkipAt || !record.review_wait_started_at) {
-    return false;
-  }
-
-  if (record.review_wait_head_sha !== pr.headRefOid) {
-    return false;
-  }
-
-  const draftSkipAtMs = Date.parse(pr.configuredBotDraftSkipAt);
-  const reviewWaitStartedAtMs = Date.parse(record.review_wait_started_at);
-  if (Number.isNaN(draftSkipAtMs) || Number.isNaN(reviewWaitStartedAtMs) || reviewWaitStartedAtMs <= draftSkipAtMs) {
-    return false;
-  }
-
-  const actionableSignalAt = latestConfiguredBotActionableSignalAt(pr);
-  if (actionableSignalAt) {
-    const actionableSignalAtMs = Date.parse(actionableSignalAt);
-    if (!Number.isNaN(actionableSignalAtMs) && actionableSignalAtMs >= reviewWaitStartedAtMs) {
-      return false;
-    }
-  }
-
-  return nowMs < reviewWaitStartedAtMs + configuredBotInitialGraceWaitMs(config);
-}
-
-function configuredBotDraftSkipRearmStartedAt(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-): string | null {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  if (!policy.shouldApplyCurrentHeadQuietPeriod || pr.isDraft || !pr.configuredBotDraftSkipAt || !record.review_wait_started_at) {
-    return null;
-  }
-
-  if (record.review_wait_head_sha !== pr.headRefOid) {
-    return null;
-  }
-
-  const draftSkipAtMs = Date.parse(pr.configuredBotDraftSkipAt);
-  const reviewWaitStartedAtMs = Date.parse(record.review_wait_started_at);
-  if (Number.isNaN(draftSkipAtMs) || Number.isNaN(reviewWaitStartedAtMs) || reviewWaitStartedAtMs <= draftSkipAtMs) {
-    return null;
-  }
-
-  const actionableSignalAt = latestConfiguredBotActionableSignalAt(pr);
-  if (actionableSignalAt) {
-    const actionableSignalAtMs = Date.parse(actionableSignalAt);
-    if (!Number.isNaN(actionableSignalAtMs) && actionableSignalAtMs >= reviewWaitStartedAtMs) {
-      return null;
-    }
-  }
-
-  return record.review_wait_started_at;
-}
-
-function shouldWaitForConfiguredBotLatestHeadRearm(
-  config: SupervisorConfig,
-  record: Pick<IssueRunRecord, "review_wait_started_at" | "review_wait_head_sha">,
-  pr: GitHubPullRequest,
-  nowMs: number,
-): boolean {
-  const startedAt = configuredBotLatestHeadRearmStartedAt(config, record, pr);
-  if (!startedAt) {
-    return false;
-  }
-
-  const startedAtMs = Date.parse(startedAt);
-  if (Number.isNaN(startedAtMs)) {
-    return false;
-  }
-
-  return nowMs < startedAtMs + configuredBotInitialGraceWaitMs(config);
-}
-
-function configuredBotLatestHeadRearmStartedAt(
-  config: SupervisorConfig,
-  record: Pick<IssueRunRecord, "review_wait_started_at" | "review_wait_head_sha">,
-  pr: GitHubPullRequest,
-): string | null {
-  const policy = reviewProviderWaitPolicyFromConfig(config);
-  if (!policy.shouldApplyCurrentHeadQuietPeriod || pr.isDraft || !record.review_wait_started_at) {
-    return null;
-  }
-
-  if (record.review_wait_head_sha !== pr.headRefOid) {
-    return null;
-  }
-
-  const reviewWaitStartedAtMs = Date.parse(record.review_wait_started_at);
-  if (Number.isNaN(reviewWaitStartedAtMs)) {
-    return null;
-  }
-
-  const actionableSignalAt = latestConfiguredBotActionableSignalAt(pr);
-  if (!actionableSignalAt) {
-    return record.review_wait_started_at;
-  }
-
-  const actionableSignalAtMs = Date.parse(actionableSignalAt);
-  if (Number.isNaN(actionableSignalAtMs)) {
-    return null;
-  }
-
-  return actionableSignalAtMs < reviewWaitStartedAtMs ? record.review_wait_started_at : null;
-}
-
-function configuredBotInitialGraceWaitMs(config: SupervisorConfig): number {
-  return (config.configuredBotInitialGraceWaitSeconds ?? DEFAULT_CONFIGURED_BOT_INITIAL_GRACE_WAIT_MS / 1_000) * 1_000;
-}
-
-function configuredBotCurrentHeadSignalWaitStartAt(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-): string | null {
-  if (
-    !requiresConfiguredBotCurrentHeadSignal(config) ||
-    pr.isDraft ||
-    hasCurrentHeadProviderSuccess(record, pr) ||
-    (validTimestamp(pr.configuredBotCurrentHeadObservedAt) && currentHeadObservationSatisfiesActiveWait(record, pr))
-  ) {
-    return null;
-  }
-
-  const currentHeadCiGreenAt = validTimestamp(pr.currentHeadCiGreenAt);
-  const codexConnectorRequiresSignal = configuredReviewProviderKinds(config).includes("codex");
-  const fallbackWaitStartedAt =
-    codexConnectorRequiresSignal && record.review_wait_head_sha === pr.headRefOid
-      ? validTimestamp(record.review_wait_started_at)
-      : codexConnectorRequiresSignal
-        ? validTimestamp(pr.createdAt)
-        : null;
-  const waitAnchor = currentHeadCiGreenAt ?? fallbackWaitStartedAt;
-  if (!waitAnchor) {
-    return null;
-  }
-
-  const waitAnchorMs = Date.parse(waitAnchor);
-  if (Number.isNaN(waitAnchorMs)) {
-    return null;
-  }
-
-  const clampWaitStartAt = (startedAt: string | null): string | null => {
-    if (!startedAt) {
-      return null;
-    }
-
-    const startedAtMs = Date.parse(startedAt);
-    if (Number.isNaN(startedAtMs)) {
-      return null;
-    }
-
-    return startedAtMs < waitAnchorMs ? waitAnchor : startedAt;
-  };
-
-  const draftSkipStartedAt = configuredBotDraftSkipRearmStartedAt(config, record, pr);
-  const clampedDraftSkipStartedAt = clampWaitStartAt(draftSkipStartedAt);
-  if (clampedDraftSkipStartedAt) {
-    return clampedDraftSkipStartedAt;
-  }
-
-  const latestHeadRearmStartedAt = configuredBotLatestHeadRearmStartedAt(config, record, pr);
-  const clampedLatestHeadRearmStartedAt = clampWaitStartAt(latestHeadRearmStartedAt);
-  if (clampedLatestHeadRearmStartedAt) {
-    return clampedLatestHeadRearmStartedAt;
-  }
-
-  return waitAnchor;
-}
-
-function configuredBotCurrentHeadSignalTimeoutStartAt(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-): string | null {
-  const waitStartedAt = configuredBotCurrentHeadSignalWaitStartAt(config, record, pr);
-  if (!waitStartedAt) {
-    return null;
-  }
-
-  const waitStartedAtMs = Date.parse(waitStartedAt);
-  if (Number.isNaN(waitStartedAtMs)) {
-    return null;
-  }
-
-  return new Date(waitStartedAtMs + configuredBotInitialGraceWaitMs(config)).toISOString();
-}
-
-function configuredBotCurrentHeadSignalPending(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-): boolean {
-  if (!requiresConfiguredBotCurrentHeadSignal(config)) {
-    return false;
-  }
-
-  return configuredBotCurrentHeadSignalWaitStartAt(config, record, pr) !== null;
+) {
+  return determineCopilotReviewTimeoutForNow(config, record, pr, nowMs);
 }
 
 export function buildCopilotReviewTimeoutFailureContext(
@@ -688,352 +133,12 @@ function mergeConditionsSatisfied(pr: GitHubPullRequest, checks: PullRequestChec
   );
 }
 
-function isIssueJournalThreadPath(thread: Pick<ReviewThread, "path">): boolean {
-  const normalizedPath = thread.path?.replace(/\\/g, "/") ?? "";
-  return /^\.codex-supervisor\/.+\/issue-journal\.md$/.test(normalizedPath);
-}
-
-function allowJournalOnlyConfiguredBotThreadException(
-  config: SupervisorConfig,
-  pr: GitHubPullRequest,
-  checks: PullRequestCheck[],
-  reviewThreads: ReviewThread[],
-): boolean {
-  if (!reviewProviderWaitPolicyFromConfig(config).shouldApplyCurrentHeadQuietPeriod) {
-    return false;
-  }
-
-  if (
-    pr.state !== "OPEN" ||
-    pr.isDraft ||
-    pr.mergedAt ||
-    pr.mergeStateStatus !== "CLEAN" ||
-    pr.mergeable !== "MERGEABLE" ||
-    pr.configuredBotCurrentHeadStatusState !== "SUCCESS" ||
-    pr.configuredBotTopLevelReviewStrength === "blocking"
-  ) {
-    return false;
-  }
-
-  const checkSummary = summarizeChecks(checks);
-  if (!checkSummary.allPassing) {
-    return false;
-  }
-
-  const unresolvedConfiguredBotThreads = configuredBotReviewThreads(config, reviewThreads);
-  return unresolvedConfiguredBotThreads.length > 0 && unresolvedConfiguredBotThreads.every(isIssueJournalThreadPath);
-}
-
-function effectiveConfiguredBotReviewThreads(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-  checks: PullRequestCheck[],
-  reviewThreads: ReviewThread[],
-): ReviewThread[] {
-  const unresolvedConfiguredBotThreads = configuredBotReviewThreads(config, reviewThreads);
-  const clearOutdatedCodexConnectorThreads = codexConnectorOutdatedThreadClearanceAllowed(
-    config,
-    record,
-    pr,
-    checks,
-    reviewThreads,
-  );
-  const effectiveThreads = codexConnectorThreadsAfterConvergencePolicy(config, pr, unresolvedConfiguredBotThreads);
-  const threadsAfterOutdatedClearance = clearOutdatedCodexConnectorThreads
-    ? effectiveThreads.filter((thread) => !isClearableOutdatedCodexConnectorResidueThread(config, thread))
-    : effectiveThreads;
-  return allowJournalOnlyConfiguredBotThreadException(config, pr, checks, threadsAfterOutdatedClearance)
-    ? threadsAfterOutdatedClearance.filter((thread) => !isIssueJournalThreadPath(thread))
-    : threadsAfterOutdatedClearance;
-}
-
-function codexConnectorThreadsAfterConvergencePolicy(
-  config: SupervisorConfig,
-  pr: GitHubPullRequest,
-  configuredThreads: ReviewThread[],
-): ReviewThread[] {
-  const codexConnectorPolicy = evaluateCodexConnectorConvergencePolicy(config, pr, configuredThreads);
-  if (codexConnectorPolicy?.outcome !== "nitpick_only" && codexConnectorPolicy?.outcome !== "converged") {
-    return configuredThreads;
-  }
-
-  const codexConnectorNitpickThreads = new Set(codexConnectorNitpickOnlyReviewThreads(configuredThreads));
-  return configuredThreads.filter((thread) => !codexConnectorNitpickThreads.has(thread));
-}
-
-function isClearableOutdatedCodexConnectorResidueThread(config: SupervisorConfig, thread: ReviewThread): boolean {
-  return (
-    thread.isOutdated &&
-    (latestReviewCommentAuthorIsAllowedBot(config, thread) || hasCodexConnectorFindingReviewComment(thread))
-  );
-}
-
-export function effectiveConfiguredBotReviewThreadsForState(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-  checks: PullRequestCheck[],
-  reviewThreads: ReviewThread[],
-): ReviewThread[] {
-  return hasProvenCodexConnectorStaleReviewMetadata({
-    config,
-    record,
-    pr,
-    checks,
-    reviewThreads,
-  })
-    ? []
-    : effectiveConfiguredBotReviewThreads(config, record, pr, checks, reviewThreads);
-}
-
-function codexConnectorOutdatedThreadClearanceAllowed(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-  checks: PullRequestCheck[],
-  reviewThreads: ReviewThread[],
-): boolean {
-  const providerKinds = configuredReviewProviderKinds(config);
-  const checkSummary = summarizeChecks(checks);
-  const configuredThreads = configuredBotReviewThreads(config, reviewThreads);
-  const codexConnectorPolicy = evaluateCodexConnectorConvergencePolicy(config, pr, configuredThreads);
-  const threadsAfterConvergencePolicy = codexConnectorThreadsAfterConvergencePolicy(config, pr, configuredThreads);
-  const onlyOutdatedCodexConnectorThreads =
-    threadsAfterConvergencePolicy.length > 0 &&
-    threadsAfterConvergencePolicy.every((thread) => isClearableOutdatedCodexConnectorResidueThread(config, thread));
-  const currentHeadCodexSuccess = hasCodexConnectorPrSuccessCurrentHeadObservation(pr);
-  const convergedOutdatedResidueCanClear =
-    providerKinds.includes("codex") &&
-    providerKinds.every((kind) => kind === "codex") &&
-    currentHeadCodexSuccess &&
-    pullRequestHeadMatchesRecord(record, pr) &&
-    checkSummary.allPassing &&
-    manualReviewThreads(config, reviewThreads).length === 0 &&
-    !mergeConflictDetected(pr) &&
-    onlyOutdatedCodexConnectorThreads &&
-    (codexConnectorPolicy?.outcome === "converged" || codexConnectorPolicy?.outcome === "nitpick_only");
-
-  return Boolean(
-    providerKinds.includes("codex") &&
-      currentHeadCodexSuccess &&
-      (currentHeadObservationSatisfiesActiveWait(record, pr) ||
-        staleSameHeadCodexWaitHasOnlyOutdatedResidue(config, record, pr, checks, reviewThreads) ||
-        convergedOutdatedResidueCanClear) &&
-      checkSummary.allPassing,
-  );
-}
-
-function staleSameHeadCodexWaitHasOnlyOutdatedResidue(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-  checks: PullRequestCheck[],
-  reviewThreads: ReviewThread[],
-): boolean {
-  const providerKinds = configuredReviewProviderKinds(config);
-  if (
-    !providerKinds.includes("codex") ||
-    providerKinds.some((kind) => kind !== "codex") ||
-    !hasCodexConnectorPrSuccessCurrentHeadObservation(pr)
-  ) {
-    return false;
-  }
-
-  if (copilotReviewPending(config, record, pr)) {
-    return false;
-  }
-
-  const observedAt = validTimestamp(pr.configuredBotCurrentHeadObservedAt);
-  const waitStartedAt = validTimestamp(record.review_wait_started_at);
-  if (!observedAt || !waitStartedAt || record.review_wait_head_sha !== pr.headRefOid) {
-    return false;
-  }
-
-  if (Date.parse(observedAt) >= Date.parse(waitStartedAt)) {
-    return false;
-  }
-
-  if (!pullRequestHeadMatchesRecord(record, pr) || mergeConflictDetected(pr) || !summarizeChecks(checks).allPassing) {
-    return false;
-  }
-
-  if (manualReviewThreads(config, reviewThreads).length > 0) {
-    return false;
-  }
-
-  const configuredThreads = configuredBotReviewThreads(config, reviewThreads);
-  const threadsAfterConvergencePolicy = codexConnectorThreadsAfterConvergencePolicy(config, pr, configuredThreads);
-  return (
-    threadsAfterConvergencePolicy.length > 0 &&
-    threadsAfterConvergencePolicy.every((thread) => isClearableOutdatedCodexConnectorResidueThread(config, thread))
-  );
-}
-
-function hasProvenCodexConnectorStaleReviewMetadata(args: {
-  config: SupervisorConfig;
-  record: IssueRunRecord;
-  pr: GitHubPullRequest;
-  checks: PullRequestCheck[];
-  reviewThreads: ReviewThread[];
-}): boolean {
-  const remediation = buildStaleReviewBotRemediation({
-    config: args.config,
-    record: args.record,
-    pr: args.pr,
-    checks: args.checks,
-    reviewThreads: args.reviewThreads,
-  });
-  return remediation ? isProvenStaleReviewMetadataClassification(remediation.classification) : false;
-}
-
-function configuredBotThreadsAllowCodexConnectorCurrentHeadWait(args: {
-  config: SupervisorConfig;
-  record: IssueRunRecord;
-  pr: GitHubPullRequest;
-  configuredThreads: ReviewThread[];
-}): boolean {
-  if (codexConnectorMustFixReviewThreads(args.configuredThreads).length > 0) {
-    return false;
-  }
-
-  if (configuredBotReviewFollowUpState(args.config, args.record, args.pr, args.configuredThreads) === "eligible") {
-    return false;
-  }
-
-  return args.configuredThreads.every(
-    (thread) =>
-      hasProcessedReviewThread(args.record, args.pr, thread) ||
-      !latestReviewCommentAuthorIsAllowedBot(args.config, thread),
-  );
-}
-
-function shouldWaitForCodexConnectorCurrentHeadReview(args: {
-  config: SupervisorConfig;
-  record: IssueRunRecord;
-  pr: GitHubPullRequest;
-  checks: PullRequestCheck[];
-  reviewThreads: ReviewThread[];
-  manualThreads: ReviewThread[];
-  unresolvedBotThreads: ReviewThread[];
-  nowMs: number;
-}): boolean {
-  if (
-    !configuredReviewProviderKinds(args.config).includes("codex") ||
-    args.pr.isDraft ||
-    mergeConflictDetected(args.pr) ||
-    args.manualThreads.length > 0 ||
-    validTimestamp(args.pr.configuredBotCurrentHeadObservedAt)
-  ) {
-    return false;
-  }
-
-  const configuredThreadsAreSafe = configuredBotThreadsAllowCodexConnectorCurrentHeadWait({
-      config: args.config,
-      record: args.record,
-      pr: args.pr,
-      configuredThreads: args.unresolvedBotThreads,
-  });
-  if (
-    codexConnectorCurrentHeadReviewReadiness({
-      config: args.config,
-      pr: args.pr,
-      checks: args.checks,
-      manualThreadCount: args.manualThreads.length,
-      configuredThreadsAreSafe,
-      checkSummary: summarizeChecks(args.checks),
-      mergeConflict: mergeConflictDetected(args.pr),
-    }).kind !== "eligible"
-  ) {
-    return false;
-  }
-
-  const timeout = determineCopilotReviewTimeout(args.config, args.record, args.pr, args.nowMs);
-  return (
-    (configuredBotCurrentHeadSignalPending(args.config, args.record, args.pr) && !timeout.timedOut) ||
-    (timeout.timedOut && timeout.action === "request_review_comment")
-  );
-}
-
-// Blocks already-processed Codex Connector must-fix threads once the tracked PR
-// repeat budget has made a no-progress stop decision.
-function processedCodexConnectorMustFixThreadsExhaustedRepeatBudget(args: {
-  config: SupervisorConfig;
-  record: IssueRunRecord;
-  pr: GitHubPullRequest;
-  checks: PullRequestCheck[];
-  manualThreads: ReviewThread[];
-  codexConnectorMustFixThreads: ReviewThread[];
-}): boolean {
-  if (
-    !configuredReviewProviderKinds(args.config).includes("codex") ||
-    args.codexConnectorMustFixThreads.length === 0 ||
-    args.manualThreads.length > 0 ||
-    args.record.last_tracked_pr_repeat_failure_decision !== "stop_no_progress"
-  ) {
-    return false;
-  }
-
-  const checkSummary = summarizeChecks(args.checks);
-  if (checkSummary.hasFailing || checkSummary.hasPending || mergeConflictDetected(args.pr)) {
-    return false;
-  }
-
-  return args.codexConnectorMustFixThreads.every((thread) =>
-    hasProcessedReviewThread(args.record, args.pr, thread),
-  );
-}
-
 function pullRequestHeadMatchesRecord(record: Pick<IssueRunRecord, "last_head_sha">, pr: GitHubPullRequest): boolean {
   return record.last_head_sha === null || record.last_head_sha === pr.headRefOid;
 }
 
 function isMergeCriticalPullRequest(pr: GitHubPullRequest): boolean {
   return pr.state === "OPEN" && !pr.isDraft && !pr.mergedAt;
-}
-
-function hasConfiguredProviderSuccess(
-  config: SupervisorConfig,
-  record: IssueRunRecord,
-  pr: GitHubPullRequest,
-  checks: PullRequestCheck[],
-  reviewThreads: ReviewThread[],
-): boolean {
-  if (!repoExpectsConfiguredBotReview(config) || !isMergeCriticalPullRequest(pr)) {
-    return false;
-  }
-
-  const codexConnectorPolicy = evaluateCodexConnectorConvergencePolicy(config, pr, reviewThreads);
-  if (codexConnectorPolicy?.outcome === "must_fix_remaining" || codexConnectorPolicy?.outcome === "missing_current_head_review") {
-    return false;
-  }
-
-  const clearOutdatedCodexConnectorThreads = codexConnectorOutdatedThreadClearanceAllowed(
-    config,
-    record,
-    pr,
-    checks,
-    reviewThreads,
-  );
-  const configuredBotThreads = configuredBotReviewThreads(config, reviewThreads).filter(
-    (thread) =>
-      !clearOutdatedCodexConnectorThreads ||
-      !isClearableOutdatedCodexConnectorResidueThread(config, thread),
-  );
-  const codexConnectorNitpickThreads = new Set(codexConnectorNitpickOnlyReviewThreads(configuredBotThreads));
-  if (configuredBotThreads.filter((thread) => !codexConnectorNitpickThreads.has(thread)).length > 0) {
-    return false;
-  }
-
-  if (pr.reviewDecision === "CHANGES_REQUESTED" && pr.configuredBotTopLevelReviewStrength !== "nitpick_only") {
-    return false;
-  }
-
-  return Boolean(
-    validTimestamp(pr.configuredBotCurrentHeadObservedAt) ||
-      validTimestamp(pr.copilotReviewArrivedAt) ||
-      (pr.configuredBotTopLevelReviewStrength === "nitpick_only" && validTimestamp(pr.configuredBotTopLevelReviewSubmittedAt)),
-  );
 }
 
 export function syncMergeLatencyVisibility(
